+++ src/scripts/index.js
@@ -42,6 +42,9 @@ import UiFormField from './components/form-field/form-field.vue';
 import UiTextfield from './components/textfield/textfield.vue';
 import UiTextfieldHelper from './components/textfield/textfield-helper.vue';
 import UiTextfieldIcon from './components/textfield/textfield-icon.vue';
+import UiAutocomplete from './components/autocomplete/autocomplete.vue'; // Custom
+import UiDatepicker from './components/datepicker/datepicker.vue'; // Custom
+import UiRangepicker from './components/rangepicker/rangepicker.vue'; // Custom
 import UiSelect from './components/select/select.vue';
 import UiSelectHelper from './components/select/select-helper.vue';
 import UiSelectIcon from './components/select/select-icon.vue';
@@ -98,6 +101,9 @@ import $store, { useStore } from './plugins/store'; // Custom
 import $grid, { useGrid } from './plugins/grid';
 import $theme, { useTheme } from './plugins/theme';
 import $typography, { useTypography } from './plugins/typography';
+import $alert, { useAlert } from './plugins/alert'; // Custom
+import $confirm, { useConfirm } from './plugins/confirm'; // Custom
+import $toast, { useToast } from './plugins/toast'; // Custom
 import $validator, { useValidator } from './plugins/validator'; // Custom
 /**
  * Directives
@@ -117,6 +123,7 @@ import helpers from './utils/helpers';
 import detectIE from './utils/ie';
 
 const Components = {
+  UiAutocomplete,
   UiBadge,
   UiBanner,
   UiButton,
@@ -131,6 +138,7 @@ const Components = {
   UiCheckbox,
   UiChip,
   UiChips,
+  UiDatepicker,
   UiDialog,
   UiDialogActions,
   UiDialogContent,
@@ -176,6 +184,7 @@ const Components = {
   UiPanel,
   UiProgress,
   UiRadio,
+  UiRangepicker,
   UiSegmentedButtons,
   UiSegmentedButton,
   UiSelect,
@@ -200,10 +209,13 @@ const Components = {
 };
 
 const Plugins = {
+  $alert,
+  $confirm,
   $event,
   $grid,
   $store,
   $theme,
+  $toast,
   $typography,
   $validator
 };
@@ -314,13 +326,26 @@ export {
   UiTooltip,
   UiTopAppBar
 };
-export { $event, $grid, $store, $theme, $typography, $validator };
 export {
+  $alert,
+  $confirm,
+  $event,
+  $grid,
+  $store,
+  $theme,
+  $toast,
+  $typography,
+  $validator
+};
+export {
+  useAlert,
+  useConfirm,
   useEvent,
   useBus,
   useGrid,
   useStore,
   useTheme,
+  useToast,
   useTypography,
   useValidator
 };