+++ src/scripts/plus.js
@@ -3,21 +3,12 @@ import autoInstall from './config/auto-install';
 /**
  * Components
  */
-import UiAutocomplete from './components/autocomplete/autocomplete.vue';
 import UiSkeleton from './components/skeleton/skeleton.vue';
-import UiDatepicker from './components/datepicker/datepicker.vue';
-import UiRangepicker from './components/rangepicker/rangepicker.vue';
 import UiAlert from './components/alert/alert.vue';
 import UiCollapse from './components/collapse/collapse.vue';
 import UiEditor from './components/editor/editor.vue';
 import UiTree from './components/tree/tree.vue';
 import UiTreeNode from './components/tree/tree-node.vue';
-/**
- * Plugins
- */
-import $alert, { useAlert } from './plugins/alert';
-import $confirm, { useConfirm } from './plugins/confirm';
-import $toast, { useToast } from './plugins/toast';
 /**
  * Directives
  */
@@ -28,21 +19,14 @@ import vDebounce from './directives/debounce';
 
 const Components = {
   UiAlert,
-  UiAutocomplete,
   UiCollapse,
-  UiDatepicker,
   UiEditor,
-  UiRangepicker,
   UiSkeleton,
   UiTree,
   UiTreeNode
 };
 
-const Plugins = {
-  $alert,
-  $confirm,
-  $toast
-};
+const Plugins = {};
 
 const Directives = {
   vAnchor,
@@ -65,17 +49,5 @@ const BalmUIPlus = {
 
 export default BalmUIPlus;
 export { version, install };
-export {
-  UiAlert,
-  UiAutocomplete,
-  UiCollapse,
-  UiDatepicker,
-  UiEditor,
-  UiRangepicker,
-  UiSkeleton,
-  UiTree,
-  UiTreeNode
-};
-export { $alert, $confirm, $toast };
-export { useAlert, useConfirm, useToast };
+export { UiAlert, UiCollapse, UiEditor, UiSkeleton, UiTree, UiTreeNode };
 export { vAnchor, vCopy, vLongpress, vDebounce };